refactor(backend): centralize auth session types
