+++ backend/src/auth/auth.service.ts
@@ -13,18 +13,9 @@ import {
   resolveRefreshJwtExpiresIn,
   resolveRefreshJwtSecret,
 } from './jwt.config';
+import type { AuthSessionPayload, SessionUser } from './auth.types';
 
-export type SessionUser = {
-  sub: string;
-  email: string;
-};
-
-export type AuthSessionPayload = {
-  user: SessionUser;
-  accessToken: string;
-  refreshToken: string;
-  csrfToken: string;
-};
+export type { AuthSessionPayload, SessionUser } from './auth.types';
 
 type RefreshTokenPayload = {
   sub: string;
