+++ backend/src/auth/auth.types.ts
@@ -1,0 +1,11 @@
+export type SessionUser = {
+  sub: string;
+  email: string;
+};
+
+export type AuthSessionPayload = {
+  user: SessionUser;
+  accessToken: string;
+  refreshToken: string;
+  csrfToken: string;
+};
